requirements.txt
# Core Data Science and Finance Libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Time Series and Volatility Modeling
arch>=6.2.0
statsmodels>=0.14.0

# Market Data and Trading APIs
yfinance>=0.2.18
alpaca-trade-api>=3.0.0
alpaca-py>=0.8.0

# Technical Analysis
ta-lib>=0.4.0
talib>=0.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and Environment
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0.0

# Database and Storage
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0

# Logging and Monitoring
loguru>=0.7.0
prometheus-client>=0.17.0

# Async and Concurrency
asyncio>=3.4.3
aiohttp>=3.8.0
websockets>=11.0.0

# Testing and Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter and Analysis
jupyter>=1.0.0
ipykernel>=6.25.0

# System and Utilities
schedule>=1.2.0
requests>=2.31.0
python-dateutil>=2.8.0
pytz>=2023.3

# Risk Management
quantlib>=1.31.0

# Machine Learning Libraries
tensorflow>=2.15.0
keras>=3.0.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Advanced Analytics
redis>=5.0.0
celery>=5.3.0
dash>=2.15.0
plotly-dash>=1.2.0

# Model Persistence
joblib>=1.3.0
dill>=0.3.7

# Performance Monitoring
psutil>=5.9.0